# frontend/public/openapi.yaml
openapi: 3.1.0
info:
  title: BugReportWeb API
  version: "1.0.0"
  summary: Bug 与功能需求跟踪、人员委派与团队 KPI 的 REST API
  description: |
    BugReportWeb 后端 REST API，用于 Bug 登记、复测、功能需求跟踪、人员委派与团队 KPI 统计。

    ## 基址与调用方式

    **请直连后端 origin**（开发默认 `http://localhost:3001`），不要以前端 Vite 开发服务器（如 `:5173`）作为 API 基址。
    生产环境若前后端同域部署，浏览器仍应使用同一 origin 下的 `/bugs`、`/auth/login` 等路径。

    请求头建议：
    - `Content-Type: application/json`（JSON 接口）
    - `Authorization: Bearer <JWT>`（除公开端点外均需要）

    ## 快速开始

    1. `POST /auth/register` 注册账号（状态为 `PENDING`）
    2. 管理员 `PATCH /users/{id}/approve` 审批并分配角色
    3. `POST /auth/login` 获取 `accessToken`（有效期 7 天）
    4. `GET /auth/me` 预检当前用户权限
    5. 按业务调用 Bug / 功能 / 系统等接口

    ```bash
    # 登录示例
    curl -X POST http://localhost:3001/auth/login \
      -H "Content-Type: application/json" \
      -d '{"username":"demo","password":"secret123"}'

    # 携带 token 创建 Bug
    curl -X POST http://localhost:3001/bugs \
      -H "Authorization: Bearer <token>" \
      -H "Content-Type: application/json" \
      -d '{"systemId":"...","title":"按钮无响应","description":"点击保存无反应","severity":"HIGH"}'
    ```

    ## 认证与账号状态

    | 用户状态 | 含义 | 能否登录 |
    |---------|------|---------|
    | `PENDING` | 已注册待审批 | 否 |
    | `ACTIVE` | 正常使用 | 是 |
    | `DISABLED` | 已禁用 | 否 |

    - JWT payload 含 `sub`（用户 ID）与 `username`
    - `isAdmin: true` 的用户**绕过所有权限检查**，但仍需有效 JWT
    - 登录失败统一返回 `401`，消息为 `Invalid credentials` 或 `User is pending approval or disabled`

    ## 权限模型

...

    ### 权限速查

...

...

...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /features:
    get:
      tags: [Features]
      summary: 功能列表
      description: |
        返回功能摘要列表。筛选语义与 Bug 列表相同（`systemId`、`status`、`deleted`、`participantUserId`）。
        详情接口才返回完整 `activities`。
      x-permissions: []
      x-service-rules:
        - deleted=only|all 时非 admin 会被拒绝或过滤
      operationId: featuresList
      parameters:
        - $ref: "#/components/parameters/systemId"
        - $ref: "#/components/parameters/featureStatus"
        - $ref: "#/components/parameters/deleted"
        - $ref: "#/components/parameters/participantUserId"
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [createdAt, plannedStart, plannedEnd, progress, title]
            default: createdAt
        - name: sortOrder
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - name: hasSchedule
          in: query
          schema:
            type: string
            enum: ['true']
      responses:
        "200":
          description: 功能列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FeatureItem"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [Features]
      summary: 创建功能
      description: |
        登记新功能需求，写入 `CREATED` 活动。初始状态为 `PLANNED`。
        必填：`systemId`、`title`、`description`。
      x-permissions: [CREATE_FEATURE]
      operationId: featuresCreate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFeatureBody"
            examples:
              default:
                value:
                  systemId: sys-001
                  title: 导出 PDF 报告
                  description: 支持按日期范围导出 Bug 统计 PDF
                  priority: MEDIUM
      responses:
        "201":
          description: 创建的功能详情
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeatureDetail"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /features/gantt:
    get:
      tags: [Features]
      summary: 功能甘特图数据
      description: |
        返回带计划时间与实现项的功能列表，默认按 `plannedStart` 升序。
        可用 `from` / `to` 过滤与时间窗口重叠的功能。
      x-permissions: []
      operationId: featuresGantt
      parameters:
        - $ref: "#/components/parameters/systemId"
        - $ref: "#/components/parameters/featureStatus"
        - $ref: "#/components/parameters/participantUserId"
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: 甘特图条目
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FeatureGanttEntry"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /features/{id}:
    get:
      tags: [Features]
      summary: 功能详情
      description: 返回完整功能含 `activities`。已软删除的仅 admin 可查看。
      x-permissions: []
      x-service-rules:
        - 已软删除的功能仅 admin 可查看
      operationId: featuresDetail
      parameters:
        - $ref: "#/components/parameters/featureId"
      responses:
        "200":
          description: 功能详情（含 activities）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeatureDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Features]
      summary: 更新功能
      description: |
        部分更新功能字段（不含 `status`）。需 `UPDATE_FEATURE`。
        非 admin 仅可编辑**自己创建**的功能。
      x-permissions: [UPDATE_FEATURE]
      x-service-rules:
        - 非 admin 仅可编辑自己创建的功能
      operationId: featuresUpdate
      parameters:
        - $ref: "#/components/parameters/featureId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFeatureBody"
      responses:
        "200":
          description: 更新后的功能
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeatureDetail"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /features/{id}/status:
    patch:
      tags: [Features]
      summary: 更新功能状态
      description: |
        切换 `PLANNED` / `IN_PROGRESS` / `DONE`，写入 `STATUS_CHANGED` 活动。
        设为 `DONE` 时记录 `completedBy` / `completedAt`。重复状态 → `400`。
      x-permissions: [UPDATE_FEATURE]
      operationId: featuresUpdateStatus
      parameters:
        - $ref: "#/components/parameters/featureId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFeatureStatusBody"
            examples:
              inProgress:
                value:
                  status: IN_PROGRESS
                  note: 已分配开发，预计本周完成
              done:
                value:
                  status: DONE
                  note: 已在 v3.0 发布
      responses:
        "200":
          description: 更新后的功能
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeatureDetail"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /features/{id}/delete:
    post:
      tags: [Features]
      summary: 软删除功能
      description: |
        软删除功能，语义同 Bug 软删除。非 admin 仅可删除自己创建的功能。
      x-permissions: [DELETE_FEATURE]
      x-service-rules:
        - 非 admin 仅可软删除自己创建的功能
      operationId: featuresSoftDelete
      parameters:
        - $ref: "#/components/parameters/featureId"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SoftDeleteBody"
      responses:
        "200":
          description: 软删除成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /features/{id}/permanent:
    delete:
      tags: [Features]
      summary: 永久删除功能
      description: |
        仅 admin；须先软删除（否则 `400 Feature must be soft-deleted first`）。不可恢复。
      x-permissions: []
      x-service-rules:
        - 仅 isAdmin 用户可调用
      operationId: featuresPermanentDelete
      parameters:
        - $ref: "#/components/parameters/featureId"
      responses:
        "200":
          description: 永久删除成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /features/{id}/activities/{activityId}:
    delete:
      tags: [Features]
      summary: 删除功能活动记录
      description: 管理用途，需 `DELETE_FEATURE_ACTIVITY`。不可作为业务撤销手段。
      x-permissions: [DELETE_FEATURE_ACTIVITY]
      operationId: featuresDeleteActivity
      parameters:
        - $ref: "#/components/parameters/featureId"
        - $ref: "#/components/parameters/activityId"
      responses:
        "200":
          description: 活动记录已删除
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /features/{id}/personnel/join:
    post:
      tags: [Features, Personnel]
      summary: 加入功能相关人
      description: |
        将当前用户加入相关人，需 `UPDATE_FEATURE`（admin 豁免）。写入 `RELATED_JOINED` 活动。
      x-permissions: [UPDATE_FEATURE]
      operationId: featuresPersonnelJoin
      parameters:
        - $ref: "#/components/parameters/featureId"
      responses:
        "200":
          description: 更新后的功能
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeatureDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /features/{id}/personnel/claim-owner:
    post:
      tags: [Features, Personnel]
      summary: 认领功能负责人
      description: 将当前用户设为 owner，需 `BECOME_ITEM_OWNER`。
      x-permissions: [BECOME_ITEM_OWNER]
      operationId: featuresPersonnelClaimOwner
      parameters:
        - $ref: "#/components/parameters/featureId"
      responses:
        "200":
          description: 更新后的功能
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeatureDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /features/{id}/personnel:
    patch:
      tags: [Features, Personnel]
      summary: 委派功能人员
      description: 动态鉴权规则与 `PATCH /bugs/{id}/personnel` 完全相同，参见该端点说明表格。
      x-permissions: []
      x-service-rules:
        - 按 PersonnelPatchBody 字段动态校验权限
      operationId: featuresPersonnelPatch
      parameters:
        - $ref: "#/components/parameters/featureId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PersonnelPatchBody"
      responses:
        "200":
          description: 更新后的功能
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeatureDetail"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /features/{id}/screenshots:
    post:
      tags: [Features, Attachments]
      summary: 上传功能截图
      description: |
        `multipart/form-data` 上传。字段 `file`（必填）须为 `image/*`，最大 **50MB**。
        写入 `SCREENSHOT_ADDED` 活动。
      x-permissions: [ADD_FEATURE_EVIDENCE]
      operationId: featuresAddScreenshot
      parameters:
        - $ref: "#/components/parameters/featureId"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                caption:
                  type: string
      responses:
        "201":
          description: 截图记录
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Screenshot"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /features/{id}/screenshots/{screenshotId}:
    delete:
      tags: [Features, Attachments]
      summary: 删除功能截图
      description: 删除截图文件并写入 `SCREENSHOT_REMOVED` 活动。
      x-permissions: [ADD_FEATURE_EVIDENCE]
      operationId: featuresRemoveScreenshot
      parameters:
        - $ref: "#/components/parameters/featureId"
        - name: screenshotId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 已删除
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /features/{id}/implementation-items:
    get:
      tags: [Features]
      summary: 实现项列表
      x-permissions: []
      operationId: featuresListImplementationItems
      parameters:
        - $ref: "#/components/parameters/featureId"
      responses:
        "200":
          description: 实现项列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ImplementationItem"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      tags: [Features]
      summary: 创建实现项
      description: 写入 `IMPLEMENTATION_ITEM_ADDED` 活动；无功能级计划时间时自动汇总子项计划。
      x-permissions: [UPDATE_FEATURE]
      operationId: featuresCreateImplementationItem
      parameters:
        - $ref: "#/components/parameters/featureId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateImplementationItemBody"
      responses:
        "201":
          description: 新建实现项
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImplementationItem"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /features/{id}/implementation-items/{itemId}:
    patch:
      tags: [Features]
      summary: 更新实现项
      x-permissions: [UPDATE_FEATURE]
      operationId: featuresUpdateImplementationItem
      parameters:
        - $ref: "#/components/parameters/featureId"
        - name: itemId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateImplementationItemBody"
      responses:
        "200":
          description: 更新后的实现项
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImplementationItem"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Features]
      summary: 删除实现项
      x-permissions: [UPDATE_FEATURE]
      operationId: featuresDeleteImplementationItem
      parameters:
        - $ref: "#/components/parameters/featureId"
        - name: itemId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 已删除
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /features/{id}/implementation-items/{itemId}/status:
    patch:
      tags: [Features]
      summary: 变更实现项状态
      description: 必填 `note`；状态前进时自动写入实际开始/完成时间。
      x-permissions: [UPDATE_FEATURE]
      operationId: featuresUpdateImplementationItemStatus
      parameters:
        - $ref: "#/components/parameters/featureId"
        - name: itemId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateImplementationItemStatusBody"
      responses:
        "200":
          description: 更新后的实现项
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImplementationItem"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 登录后获得的 accessToken，有效期 7 天。
  parameters:
...
    featureId:
      name: id
      in: path
      required: true
      description: 功能 ID（cuid）
      schema:
        type: string
        example: clxfeat000001
    activityId:
      name: activityId
      in: path
      required: true
      description: 活动记录 ID
      schema:
        type: string
...
    systemId:
      name: systemId
      in: query
      description: 按被跟踪系统 ID 过滤列表
      schema:
        type: string
...
    featureStatus:
      name: status
      in: query
      description: 按功能状态过滤
      schema:
        $ref: "#/components/schemas/FeatureStatus"
    deleted:
      name: deleted
      in: query
      description: |
        软删除过滤：`active`（默认）、`only`（回收站，非 admin 403）、`all`（全部）
      schema:
        type: string
        enum: [active, only, all]
        default: active
    participantUserId:
      name: participantUserId
      in: query
      description: 筛选指定用户参与过的 Bug/功能（创建者、负责人、相关人等）
      schema:
        type: string
...
  responses:
    BadRequest:
      description: 请求参数无效或业务校验失败
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            validation:
              value:
                statusCode: 400
                message: systemId, title and description are required
                error: Bad Request
    Unauthorized:
      description: 未认证、token 无效、或账号未激活
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            invalidToken:
              value:
                statusCode: 401
                message: Unauthorized
    Forbidden:
      description: 权限不足或违反服务层访问规则
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            noPermission:
              value:
                statusCode: 403
                message: You do not have permission to delegate the owner
                error: Forbidden
    NotFound:
      description: 资源不存在或当前用户无权查看（如已软删除）
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            missing:
              value:
                statusCode: 404
                message: Bug not found
                error: Not Found
...
  schemas:
...
    OkResponse:
      type: object
      description: 无 body 业务数据时的成功响应
      required: [ok]
      properties:
        ok:
          type: boolean
          const: true
    SoftDeleteBody:
      type: object
      description: 软删除可选原因
      properties:
        reason:
          type: string
          description: 删除原因，展示于 deleteReason 字段
...
    UpdateFeatureStatusBody:
      type: object
      required: [status]
      properties:
        status:
          $ref: "#/components/schemas/FeatureStatus"
        note:
          type: string
          description: 状态变更备注（可选）
...
    FeatureStatus:
      type: string
      description: "功能状态：PLANNED（计划中）、IN_PROGRESS（进行中）、DONE（已完成）"
      enum: [PLANNED, IN_PROGRESS, DONE]
...
    PersonnelPatchBody:
      type: object
      description: 至少提供一个字段；动态鉴权规则见 PATCH /bugs|features/{id}/personnel
      properties:
        ownerId:
          type: string
          nullable: true
          description: 委派负责人用户 ID；传 null 撤销负责人
        addRelatedUserIds:
          type: array
          description: 要添加的相关人用户 ID 列表（须 ACTIVE，且不能含 owner）
          items:
            type: string
        removeRelatedUserIds:
          type: array
          description: 要移除的相关人 ID；传自己的 ID 可自助退出
          items:
            type: string
    Screenshot:
      type: object
      properties:
        id:
          type: string
        path:
          type: string
          description: 相对路径，完整 URL 为 `{API_BASE}{path}` 或 `{API_BASE}/uploads/...`
        originalName:
          type: string
        caption:
          type: string
          nullable: true
        uploaderId:
          type: string
...
    FeatureItem:
      type: object
      description: 功能列表项或详情主体（详情另含 activities）
      properties:
        id:
          type: string
          description: 功能唯一 ID（cuid）
        title:
          type: string
          description: 功能标题
        description:
          type: string
          description: 功能描述与验收标准
        status:
          $ref: "#/components/schemas/FeatureStatus"
        priority:
          $ref: "#/components/schemas/Severity"
        system:
          $ref: "#/components/schemas/TrackedSystem"
        creator:
          $ref: "#/components/schemas/AssignableUser"
        owner:
          oneOf:
            - $ref: "#/components/schemas/AssignableUser"
            - type: "null"
        relatedUsers:
          type: array
          items:
            $ref: "#/components/schemas/AssignableUser"
        completedBy:
          oneOf:
            - $ref: "#/components/schemas/AssignableUser"
            - type: "null"
        completedAt:
          type: string
          format: date-time
          nullable: true
        plannedStartAt:
          type: string
          format: date-time
          nullable: true
        plannedEndAt:
          type: string
          format: date-time
          nullable: true
        effectivePlannedStartAt:
          type: string
          format: date-time
          nullable: true
        effectivePlannedEndAt:
          type: string
          format: date-time
          nullable: true
        progressPercent:
          type: integer
          nullable: true
        implementationItemCount:
          type: integer
        implementationItemDoneCount:
          type: integer
        screenshotCount:
          type: integer
        deletedBy:
          oneOf:
            - $ref: "#/components/schemas/AssignableUser"
            - type: "null"
        deletedAt:
          type: string
          format: date-time
          nullable: true
        deleteReason:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    FeatureDetail:
      allOf:
        - $ref: "#/components/schemas/FeatureItem"
        - type: object
          properties:
            screenshots:
              type: array
              items:
                $ref: "#/components/schemas/Screenshot"
            implementationItems:
              type: array
              items:
                $ref: "#/components/schemas/ImplementationItem"
            activities:
              type: array
              items:
                $ref: "#/components/schemas/FeatureActivity"
    FeatureGanttEntry:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        status:
          $ref: "#/components/schemas/FeatureStatus"
        system:
          $ref: "#/components/schemas/TrackedSystem"
        owner:
          oneOf:
            - $ref: "#/components/schemas/AssignableUser"
            - type: "null"
        plannedStartAt:
          type: string
          format: date-time
          nullable: true
        plannedEndAt:
          type: string
          format: date-time
          nullable: true
        effectivePlannedStartAt:
          type: string
          format: date-time
          nullable: true
        effectivePlannedEndAt:
          type: string
          format: date-time
          nullable: true
        progressPercent:
          type: integer
          nullable: true
        implementationItemCount:
          type: integer
        implementationItemDoneCount:
          type: integer
        implementationItems:
          type: array
          items:
            $ref: "#/components/schemas/ImplementationItem"
...
    UpdateImplementationItemStatusBody:
      type: object
      required: [status, note]
      properties:
        status:
          $ref: "#/components/schemas/ImplementationItemStatus"
        note:
          type: string
...
    CreateFeatureBody:
      type: object
      required: [systemId, title, description]
      properties:
        systemId:
          type: string
          description: 关联系统 ID
        title:
          type: string
          description: 功能标题
        description:
          type: string
          description: 功能描述与验收标准
        priority:
          $ref: "#/components/schemas/Severity"
        plannedStartAt:
          type: string
          format: date-time
          nullable: true
        plannedEndAt:
          type: string
          format: date-time
          nullable: true
    UpdateFeatureBody:
      type: object
      properties:
        systemId:
          type: string
        title:
          type: string
        description:
          type: string
        priority:
          $ref: "#/components/schemas/Severity"
        plannedStartAt:
          type: string
          format: date-time
          nullable: true
        plannedEndAt:
          type: string
          format: date-time
          nullable: true
...